take comments
