+++ src/kili/services/types.py
@@ -1,9 +1,9 @@
 """
 Types used by the conversion service
 """
-from typing import NamedTuple, NewType, Tuple, Union
+from typing import Any, Dict, List, NamedTuple, NewType, Tuple, Union
 
-from typing_extensions import Literal
+from typing_extensions import Literal, TypedDict
 
 InputType = Literal["TEXT", "IMAGE"]
 AssetId = NewType("AssetId", str)
@@ -24,3 +24,28 @@ class JobCategory(NamedTuple):
 
 YoloAnnotation = Tuple[int, float, float, float, float]
 LabelType = Literal["AUTOSAVE", "DEFAULT", "PREDICTION", "REVIEW"]
+
+InputType = Literal["IMAGE", "TEXT"]
+MLTask = Literal["CLASSIFICATION", "NAMED_ENTITIES_RECOGNITION", "OBJECT_DETECTION"]
+Tool = Literal["rectangle", "semantic", "polygon"]
+JobName = NewType("JobName", str)
+ProjectId = NewType("ProjectId", str)
+
+
+class Job(TypedDict):
+    """
+    Contains job settings
+    """
+
+    content: Any
+    instruction: str
+    isChild: bool
+    tools: List[Tool]
+    mlTask: MLTask
+    models: Any  # example: {"interactive-segmentation": {"job": "SEMANTIC_JOB_MARKER"}},
+    isVisible: bool
+    required: int
+    isNew: bool
+
+
+Jobs = Dict[JobName, Job]
